+++ allennlp/tests/training/learning_rate_schedulers/learning_rate_scheduler_test.py
@@ -1,0 +1,41 @@
+# pylint: disable=no-self-use,invalid-name,protected-access
+
+import torch
+
+from allennlp.common.checks import ConfigurationError
+from allennlp.training.optimizers import Optimizer
+from allennlp.common.testing import AllenNlpTestCase
+from allennlp.training.learning_rate_schedulers import LearningRateScheduler
+from allennlp.common.params import Params
+
+
+class LearningRateSchedulersTest(AllenNlpTestCase):
+
+    def test_reduce_on_plateau_error_throw_when_no_metrics_exist(self):
+        model = torch.nn.Sequential(torch.nn.Linear(10, 10))
+        with self.assertRaises(ConfigurationError) as context:
+            LearningRateScheduler.from_params(Optimizer.from_params(model.named_parameters(),
+                                                                    Params({"type": "adam"})),
+                                              Params({"type": "reduce_on_plateau"})).step(None, None)
+        assert "learning rate scheduler requires a validation metric" in str(context.exception)
+
+    def test_reduce_on_plateau_works_when_metrics_exist(self):
+        model = torch.nn.Sequential(torch.nn.Linear(10, 10))
+        LearningRateScheduler.from_params(Optimizer.from_params(model.named_parameters(),
+                                                                Params({"type": "adam"})),
+                                          Params({"type": "reduce_on_plateau"})).step(10, None)
+
+    def test_no_metric_wrapper_can_support_none_for_metrics(self):
+        model = torch.nn.Sequential(torch.nn.Linear(10, 10))
+        lrs = LearningRateScheduler.from_params(Optimizer.from_params(model.named_parameters(),
+                                                                      Params({"type": "adam"})),
+                                                Params({"type": "step", "step_size": 1}))
+        lrs.step(None, None)
+
+    def test_noam_learning_rate_schedule_does_not_crash(self):
+        model = torch.nn.Sequential(torch.nn.Linear(10, 10))
+        lrs = LearningRateScheduler.from_params(Optimizer.from_params(model.named_parameters(),
+                                                                      Params({"type": "adam"})),
+                                                Params({"type": "noam", "model_size": 10, "warmup_steps": 2000}))
+        lrs.step(None)
+        lrs.step_batch(None)
